Polygon collision performance (#173)
Polygon collision performance regressed from v2.4 due to the addition of
speculation. I've rewritten the algorithm to reduce the cost especially
for the small overlap case which is common.

Polygon collision performance is about 35% faster and now performs
similarly to v2.4 when single threaded while keeping the new features
and supporting multi-threading.